expression: fix 2 unstable test cases
